more consolidation

+++ clap_slice/audio_ordering_candidates_registry.py
@@ -5,11 +5,28 @@ from clap_slice.audio_orderer import AudioOrdering
 
 
 def save_candidate(audio_ordering: AudioOrdering, registry_root='./audio_ordering_candidates_registry'):
+    if already_saved(audio_ordering, registry_root):
+        print('already saved')
+        return None
     source_subfolder = make_source_subfolder(registry_root, source_audio_path=audio_ordering.source_audio)
     candidate_index = get_next_candidate_index(source_subfolder)
     os.makedirs(source_subfolder, exist_ok=True)
-    with open(make_candidate_path(source_subfolder, candidate_index), "wb") as f:
+    path = make_candidate_path(source_subfolder, candidate_index)
+    with open(path, "wb") as f:
         pickle.dump(audio_ordering, f)
+    return path
+
+
+def already_saved(audio_ordering: AudioOrdering, registry_root: str) -> bool:
+    source_subfolder = make_source_subfolder(registry_root, source_audio_path=audio_ordering.source_audio)
+    final_index = get_next_candidate_index(source_subfolder)
+    for i in range(final_index):
+        with open(make_candidate_path(source_subfolder, i), "rb") as f:
+            so = pickle.load(f)
+            if audio_ordering.__dict__ == so.__dict__:
+                return True
+    return False
+
 
 
 def get_next_candidate_index(source_subfolder: str) -> int:
